pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "bookstore API"
authors = [
    {name = "Luan-LopS",email = "luam@hotmail.com.br"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "factory-boy (==3.2.0)",
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pytest = ">=7.0"
pytest-django = ">=4.11.1"
black = "^25.12.0"

